+++ src/utils/AuthUtils.ts
@@ -2,8 +2,7 @@ import { Session } from 'next-auth';
 
 const AuthUtils = {
 	validateLoginSession: (session: Session | null, isLoading?: boolean) => {
-		console.log('validating login session');
-		isLoading = (isLoading != null) ? isLoading : false;
+				isLoading = (isLoading != null) ? isLoading : false;
 
 		if (isLoading) {
 			return null;
@@ -12,7 +11,6 @@ const AuthUtils = {
 		if (!isLoading && !session) {
 			window.open('/api/auth/signin', '_self');
 		}
-		console.log('session validated');
 	},
 };