+++ src/utils/MongoDBUtils.ts
@@ -1,14 +1,26 @@
-import { MongoClient, MongoClientOptions } from 'mongodb';
+import { Db, MongoClient, MongoClientOptions } from 'mongodb';
 import constants from '../constants/constants';
 
-const MongoDBUtils = async (database: string): Promise<MongoClient> => {
-	const options: MongoClientOptions = {
-		writeConcern: {
-			w: 'majority',
-		},
-		useUnifiedTopology: true,
-	};
-	return MongoClient.connect(constants.MONGODB_URI_PARTIAL + database, options);
+const MongoDBUtils = {
+	state: {
+		dbMap: new Map<string, Db>(),
+		clientMap: new Map<string, MongoClient>(),
+	},
+
+	connectDb: async (database: string): Promise<Db> => {
+		const db: Db | undefined = MongoDBUtils.state.dbMap.get(database);
+		if (db == null) {
+			const options: MongoClientOptions = {
+				writeConcern: {
+					w: 'majority',
+				},
+				useUnifiedTopology: true,
+			};
+			return (await MongoClient.connect(constants.MONGODB_URI_PARTIAL + database, options)).db();
+		} else {
+			return db;
+		}
+	},
 };
 
 export default MongoDBUtils;