+++ CHANGELOG.md
@@ -3,5 +3,6 @@
 ## 1.0.0-SNAPSHOT
 
 1. Add initial nextjs setup
-2. Integrate next-auth
-3. Integrate discord auth
+2. Integration discord auth
+   - integrate twitter auth
+   - link accounts